♻️ ai: 清理已迁移至 cago 的残留代码和测试
- 删除 opsctl_policy_test、app_ai_render_test 等已迁移测试
- 移除 conversation entity/repo/service 中旧的 session 列引用
- 精简 kafka_helper、command_policy 中不再需要的逻辑
- 提取 policy_test_helpers 供复用
