+++ internal/ai/command_policy.go
@@ -164,10 +164,6 @@ func grantItemMatchesTarget(item *grant_entity.GrantItem, assetID int64, groupID
 	return true
 }
 
-// Reset 重置会话级白名单（已迁移到 DB Grant，无需内存清理）
-func (c *CommandPolicyChecker) Reset() {
-}
-
 // Check 检查命令是否允许执行
 func (c *CommandPolicyChecker) Check(ctx context.Context, assetID int64, command string) CheckResult {
 	result := CheckPermission(ctx, asset_entity.AssetTypeSSH, assetID, command)
@@ -175,27 +171,6 @@ func (c *CommandPolicyChecker) Check(ctx context.Context, assetID int64, command
 		return result
 	}
 	return c.handleConfirm(ctx, assetID, asset_entity.AssetTypeSSH, command)
-}
-
-// CheckPolicyOnly 只检查 allow/deny 列表 + DB Grant 匹配，不触发确认回调。
-// 向后兼容包装器，内部委托 CheckPermission。
-func CheckPolicyOnly(ctx context.Context, assetID int64, command string) CheckResult {
-	return CheckPermission(ctx, asset_entity.AssetTypeSSH, assetID, command)
-}
-
-// CheckSQLPolicyForOpsctl 检查 SQL 策略，向后兼容包装器，内部委托 CheckPermission。
-func CheckSQLPolicyForOpsctl(ctx context.Context, assetID int64, sqlText string) CheckResult {
-	return CheckPermission(ctx, asset_entity.AssetTypeDatabase, assetID, sqlText)
-}
-
-// CheckRedisPolicyForOpsctl 检查 Redis 策略，向后兼容包装器，内部委托 CheckPermission。
-func CheckRedisPolicyForOpsctl(ctx context.Context, assetID int64, command string) CheckResult {
-	return CheckPermission(ctx, asset_entity.AssetTypeRedis, assetID, command)
-}
-
-// CheckKafkaPolicyForOpsctl 检查 Kafka 策略，内部委托 CheckPermission。
-func CheckKafkaPolicyForOpsctl(ctx context.Context, assetID int64, command string) CheckResult {
-	return CheckPermission(ctx, asset_entity.AssetTypeKafka, assetID, command)
 }
 
 // CheckForAsset 按资产类型分发权限检查
