+++ internal/model/entity/conversation_entity/conversation.go
@@ -2,7 +2,6 @@ package conversation_entity
 
 import (
 	"encoding/json"
-	"fmt"
 )
 
 // 状态常量
@@ -23,40 +22,11 @@ type Conversation struct {
 	Status       int    `gorm:"column:status;default:1"`
 	Createtime   int64  `gorm:"column:createtime"`
 	Updatetime   int64  `gorm:"column:updatetime"`
-	// cago State 平铺字段（202605080010 迁移之后写入；gormStore 单源）
-	ThreadID    string `gorm:"column:thread_id;type:varchar(255)"`
-	StateValues string `gorm:"column:state_values;type:text"`
 }
 
 // TableName GORM表名
 func (Conversation) TableName() string {
 	return "conversations"
-}
-
-// GetStateValues 反序列化 cago Session State.Values。
-func (c *Conversation) GetStateValues() (map[string]string, error) {
-	if c.StateValues == "" {
-		return nil, nil
-	}
-	var m map[string]string
-	if err := json.Unmarshal([]byte(c.StateValues), &m); err != nil {
-		return nil, fmt.Errorf("解析 state_values 失败: %w", err)
-	}
-	return m, nil
-}
-
-// SetStateValues 序列化 cago Session State.Values；nil/空 map 视为清空。
-func (c *Conversation) SetStateValues(v map[string]string) error {
-	if len(v) == 0 {
-		c.StateValues = ""
-		return nil
-	}
-	data, err := json.Marshal(v)
-	if err != nil {
-		return fmt.Errorf("序列化 state_values 失败: %w", err)
-	}
-	c.StateValues = string(data)
-	return nil
 }
 
 // Message 会话消息实体（cago agents v2 schema）
